+++ src/cli/mod.rs
@@ -1,0 +1,7 @@
+pub mod build;
+pub mod clean;
+pub mod flag_update;
+pub mod help;
+pub mod init;
+pub mod new;
+pub mod run;